requirements.txt
pandas==1.5.3
tqdm==4.64.0
langchain==0.0.146
langchain-google-genai==0.0.4
gnews==0.3.6
bing-image-downloader==1.1.2
Pillow==9.2.0
requests==2.28.1
streamlit==1.33.0